refactor: AuctionRepository 패키지 이동
infra -> domain
